Move reassignment to the ty module.
