+++ sway-core/src/language/ty/declaration/mod.rs
@@ -1,4 +1,6 @@
 #[allow(clippy::module_inception)]
 mod declaration;
+mod reassignment;
 
 pub use declaration::*;
+pub use reassignment::*;
